backend/requirements.txt
Flask
flask-cors
GitPython
python-dotenv
requests
gunicorn
